feat: allow bot to add role to logged in clients
The bot will assign the role "Account linked" role to users upon successful login.
The role will give users access to ranking commands (not yet implemented), and prevent users from calling /login command again. This prevents duplicate entries in the DB, on top of the existing measures to discourage this.
The bot MUST have the permission "Manage Roles" in the guild, and the role of the bot must be place higher than the role(s) the bot will be assigning to users in the hierarchy.