feat: add option to hard reset user to reset commands
Add optional parameter to the admin command ResetMember.
If the param hard is true, then the all data on the user would be removed, and the user would lose the verified role.
This option is useful if member A has somehow found a way to verify their discord account using another member B's email. Member B would then not be able to verify their own account. This option allows admins to hard reset member A's account.